Refactor app.py for background removal functionality

+++ app.py
@@ -40,7 +40,7 @@ device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
 print(f"Using device: {device}")
 
 model.loadcheckpoints(model_path)
-model.to(device)
+model.to(device).float()
 model.eval()
 
 output_folder = 'output_images'
